dag menu config and toolbar config
